# storage/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlcgen

import (
	"encoding/json"
)

type HyperliquidStatusHistory struct {
	ID            int64  `json:"id"`
	Md            string `json:"md"`
	Status        []byte `json:"status"`
	RecordedAtUtc int64  `json:"recorded_at_utc"`
}

type HyperliquidSubmission struct {
	Md             string          `json:"md"`
	ActionKind     string          `json:"action_kind"`
	CreatePayload  json.RawMessage `json:"create_payload"`
	ModifyPayloads json.RawMessage `json:"modify_payloads"`
	CancelPayload  json.RawMessage `json:"cancel_payload"`
	UpdatedAtUtc   int64           `json:"updated_at_utc"`
	BoteventRowID  int64           `json:"botevent_row_id"`
}

type KvState struct {
	K            string `json:"k"`
	V            []byte `json:"v"`
	UpdatedAtUtc int64  `json:"updated_at_utc"`
}

type ThreecommasBot struct {
	BotID         int64  `json:"bot_id"`
	Payload       []byte `json:"payload"`
	LastSyncedUtc int64  `json:"last_synced_utc"`
}

type ThreecommasBotevent struct {
	ID            int64  `json:"id"`
	Md            string `json:"md"`
	BotID         int64  `json:"bot_id"`
	DealID        int64  `json:"deal_id"`
	BoteventID    int64  `json:"botevent_id"`
	CreatedAtUtc  int64  `json:"created_at_utc"`
	ObservedAtUtc int64  `json:"observed_at_utc"`
	Payload       []byte `json:"payload"`
}

type ThreecommasBoteventsLog struct {
	ID            int64  `json:"id"`
	Md            string `json:"md"`
	BotID         int64  `json:"bot_id"`
	DealID        int64  `json:"deal_id"`
	BoteventID    int64  `json:"botevent_id"`
	CreatedAtUtc  int64  `json:"created_at_utc"`
	ObservedAtUtc int64  `json:"observed_at_utc"`
	Payload       []byte `json:"payload"`
}

type ThreecommasDeal struct {
	DealID        int64  `json:"deal_id"`
	BotID         int64  `json:"bot_id"`
	CreatedAtUtc  int64  `json:"created_at_utc"`
	UpdatedAtUtc  int64  `json:"updated_at_utc"`
	Payload       []byte `json:"payload"`
	InsertedAtUtc int64  `json:"inserted_at_utc"`
}

type VaultPayload struct {
	ID             int64           `json:"id"`
	UserID         int64           `json:"user_id"`
	Version        string          `json:"version"`
	Ciphertext     []byte          `json:"ciphertext"`
	Nonce          []byte          `json:"nonce"`
	AssociatedData []byte          `json:"associated_data"`
	PrfParams      json.RawMessage `json:"prf_params"`
	UpdatedAtUtc   int64           `json:"updated_at_utc"`
}

type VaultUser struct {
	ID           int64  `json:"id"`
	Username     string `json:"username"`
	CreatedAtUtc int64  `json:"created_at_utc"`
}

type WebauthnCredential struct {
	ID           int64           `json:"id"`
	UserID       int64           `json:"user_id"`
	CredentialID []byte          `json:"credential_id"`
	Credential   json.RawMessage `json:"credential"`
	CreatedAtUtc int64           `json:"created_at_utc"`
	UpdatedAtUtc int64           `json:"updated_at_utc"`
}